+++ src/bin/am/commands.rs
@@ -1,4 +1,3 @@
-use crate::updater;
 use anyhow::Result;
 use autometrics_am::config::AmConfig;
 use clap::{Parser, Subcommand};
@@ -9,6 +8,7 @@ use tracing::info;
 mod explore;
 pub mod start;
 pub mod system;
+pub mod update;
 
 #[derive(Parser)]
 #[command(author, version, about, long_about = None, display_name = "am")]
@@ -49,7 +49,7 @@ pub enum SubCommands {
     Discord,
 
     /// Run the updater
-    Update(updater::Arguments),
+    Update(update::Arguments),
 
     #[clap(hide = true)]
     MarkdownHelp,
@@ -69,7 +69,7 @@ pub async fn handle_command(app: Application, config: AmConfig, mp: MultiProgres
 
             Ok(())
         }
-        SubCommands::Update(args) => updater::handle_command(args, mp).await,
+        SubCommands::Update(args) => update::handle_command(args, mp).await,
         SubCommands::MarkdownHelp => {
             clap_markdown::print_help_markdown::<Application>();
             Ok(())
